requirements.txt
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1
django-cprofile-middleware==1.0.4
django-import-export==1.0.1
django-widget-tweaks==1.4.2
docxtpl==0.5.0
et-xmlfile==1.0.1
jdcal==1.4
Jinja2==2.10
lxml==4.2.4
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.5
pipdeptree==0.13.0
pyodbc==4.0.23
python-decouple==3.1
python-docx==0.8.6
pytz==2018.5
PyYAML==3.13
selenium==3.14.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
